substantial reorg and simplification
git-svn-id: http://biofs04/svn/bioinf/anfo/branches/alt-aligner@2349 93d9a472-daf7-0310-8f21-d7befa554562

+++ src/index.cc
@@ -255,7 +255,7 @@ unsigned FixedIndex::lookupS( const std::string& dna, PreSeeds& v,
 	unsigned filled = 0 ;
 	unsigned total = 0 ;
 
-	for( int32_t offset = 0 ; offset != dna.length() ; ++offset )
+	for( int32_t offset = 0 ; offset != (int32_t)dna.length() ; ++offset )
 	{
 		o_f >>= 2 ;
 		o_r = (o_r << 2) & mask ;
